+++ example/data-persistence-tf/variables.tf
@@ -99,15 +99,17 @@ variable "dbRecreation" {
 
 variable "db_partition_config" {
   type = object({
-    executions_total_years = number
-    granules_count         = number
-    files_count            = number
+    executions_total_years     = number
+    granules_count             = number
+    files_count                = number
+    executions_retention_years = number
   })
   description = "Partition configuration for this deployment"
   default = {
-    executions_total_years = 3
-    granules_count         = 4
-    files_count            = 4
+    executions_total_years     = 3
+    granules_count             = 4
+    files_count                = 4
+    executions_retention_years = 2
   }
 }
 variable "use_bootstrap" {