+++ lambdas/db-migration/variables.tf
@@ -39,9 +39,10 @@ variable "vpc_id" {
 
 variable "db_partition_config" {
   type = object({
-    executions_total_years = number
-    granules_count         = number
-    files_count            = number
+    executions_total_years     = number
+    granules_count             = number
+    files_count                = number
+    executions_retention_years = number
   })
   description = "Partitioning settings for the database migration Lambda"
 }